requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiosqlite==0.19.0
albumentations==1.3.0
ale-py==0.7.4
altair==5.0.1
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
async-timeout==4.0.2
attr==0.3.1
attrs==23.1.0
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
azure-core==1.26.4
azure-storage-blob==12.16.0
Babel==2.12.1
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
beautifulsoup4==4.12.2
bleach==5.0.1
blinker==1.6.2
boto==2.49.0
boto3==1.16.63
botocore==1.19.63
boxing==0.1.4
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==7.1.2
cloudpickle==2.2.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1679481329611/work
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
cycler==0.11.0
debugpy @ file:///D:/bld/debugpy_1680755658559/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
defusedxml==0.7.1
Deprecated==1.2.14
Django==3.2.16
django-annoying==0.10.6
django-cors-headers==3.6.0
django-debug-toolbar==3.2.1
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rest-swagger==2.2.0
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
drf-yasg==1.20.0
easyocr==1.6.2
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
expiringdict==1.2.2
fastjsonschema==2.16.3
filelock==3.9.0
fonttools==4.39.3
fqdn==1.5.1
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.10.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.3.2
google-cloud-logging==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.3
grpcio==1.54.0
grpcio-status==1.48.2
gym==0.21.0
h11==0.14.0
htmlmin==0.1.12
huggingface-hub==0.13.4
idna==3.4
imageio==2.27.0
importlib-metadata==4.13.0
importlib-resources==5.12.0
inflection==0.5.1
ipykernel @ file:///D:/bld/ipykernel_1679336472934/work
ipython @ file:///D:/bld/ipython_1682783526100/work
ipython-genutils==0.2.0
isodate==0.6.1
isoduration==20.11.0
itypes==1.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-server==1.24.0
jupyter-ydoc==0.2.4
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1681432441054/work
jupyter_core @ file:///D:/bld/jupyter_core_1678994368923/work
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.0.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.22.1
kiwisolver==1.4.4
label-studio==1.7.3
label-studio-converter==0.0.51
label-studio-tools==0.0.2
launchdarkly-server-sdk==7.5.0
lazy_loader==0.2
lockfile==0.12.2
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdurl==0.1.2
mistune==2.0.5
mpmath==1.2.1
mss==9.0.1
nbclassic==0.3.7
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.4.3
networkx==3.0
ninja==1.11.1
nltk==3.6.7
notebook==6.4.5
notebook-utils==0.2.0
notebook_shim==0.2.3
numpy==1.21.6
oauthlib==3.2.2
openapi-codec==1.3.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opencv-python-headless==4.5.4.60
openvino==2023.0.0
ordered-set==4.0.2
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==1.3.5
pandocfilters==1.5.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1682644429438/work
prometheus-client==0.16.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1677600924538/work
proto-plus==1.22.2
protobuf==3.20.3
psutil @ file:///D:/bld/psutil_1681775298734/work
psycopg2-binary==2.9.6
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post4
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.1b0
PyDirectInput==1.0.4
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1681904169130/work
Pympler==1.0.1
pyngrok==6.0.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.19.3
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==305.1
pywinpty==2.0.10
PyYAML==6.0
pyzmq @ file:///D:/bld/pyzmq_1679317071559/work
qudida==0.0.4
redis==4.5.4
regex==2023.3.23
requests==2.28.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.1
rq==1.10.1
rsa==4.9
ruamel.yaml==0.17.22
ruamel.yaml.clib==0.2.7
rules==2.2
s3transfer==0.3.7
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
semver==2.13.0
Send2Trash==1.8.2
sentencepiece==0.1.98
sentry-sdk==1.21.1
shapely==2.0.1
simplejson==3.19.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
stable-baselines3==1.8.0
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
streamlit==1.23.1
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
torchmetrics==0.11.4
torchvision==0.15.1+cu117
tornado @ file:///D:/bld/tornado_1681817629351/work
tqdm==4.65.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
transformers==4.28.1
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3
ua-parser==0.16.1
ujson==5.7.0
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.13
user-agents==2.2.0
utils==1.0.1
uvicorn==0.13.1
validators==0.20.0
watchdog==3.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.4
wget==3.2
wrapt==1.15.0
xmljson==0.2.0
y-py==0.5.9
ypy-websocket==0.8.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1677313463193/work
